chore: update README.md to better explain repo
I have completed several projects for the nanodegree program so far, many of which have been done on their workspace. I plan to download them and add them to this repo so employers can see my work.

Added the types of projects that will be shown to give an idea of what I have accomplished so far and which languages were used.